support env update
